requirements.txt
pytube
customtkinter
Pillow
qrcode
pyperclip